Make sure staging GUID is consistent
For some reason the staging guid appears in two places:
- as a variable in the staging container environment
- as a field in the staging task model

There is only one staging guid, so the env var should always be the same as the
model field. This commit reflects that in the tests and makes sure we
use the opi.StagingTask.StagingGUID as the k8s job name.

[#171785772]

Signed-off-by: Julian Skupnjak <skupnjak@de.ibm.com>
Co-authored-by: Julian Skupnjak <skupnjak@de.ibm.com>