Allow AutomountServiceAccountToken in pods via config map setting
Introduce the `UnsafeAllowAutomountServiceAccountToken` config map
property.

When unset, or set to false, the existing behaviour of statefulsets
setting `automountServiceAccountToken: &false` for pods will continue.
By setting the new property to true, `automountServiceAccountToken` will
not be set in the statefulset pod spec, allowing the value from the
ServiceAccount used for app deployment to be used.

[#173403187]

Signed-off-by: Danail Branekov <danailster@gmail.com>