Refactor eats test suite to run tests on deployed eirini
When running on a deployed eirini we do not need to run local binaries.
Instead, we are requesting the endpoint of the opi that has been already
deployed

The `USE_DEPLOYED_EIRINI` env variable determines whether we are running
agains the deployed eirini or not.

There are 2 tests still failing when you use USE_DEPLOYED_EIRINI because
they rely on a fake cloud controller being started
(Might be fixed with this: https://www.pivotaltracker.com/story/show/174193327).

Signed-off-by: Mario Nitchev <marionitchev@gmail.com>