Run integration tests in a Pod
The new script creates a pod and the relevant rbac objects, copies the
code in the pod and runs the integration tests.

The logs are streamed and the exit code of the script is the exit code of
the tests.

The script cleans up the created pod unless SKIP_CLEANUP is set to
something != "false".

The generated pod name is unique so one can run the script multiple
times without replacing or deleting older runs.

[#174354675]

Co-authored-by: Georgi Dankov <gddankov@gmail.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Dimitris Karakasilis <dkarakasilis@suse.com>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>