Remove compilation of opi binary from eats suite
All eats are now running only against a standalone eirini deployment, so
there is no need to specify USE_DEPLOYED_EIRINI or HELMLESS_DEPLOYMENT
env vars.

Since some eirini binaries are not part of the standalone deployment yet
their respective tests are still compiling and running the respective
binary on the local machine. As we introduce more binaries to the
standalone deployment we should adapt those tests to use the deployment
only

Co-authored-by: Mario Nitchev <marionitchev@gmail.com>

[#174363619]