Task reporter does not depend on previous pod state
K8S informers receive previous and current state of the objects they
monitor, however it seems that it is not guaranteed that the previous
state is the state immediately before the current one. This means that
implementing the informer to only send completion notification when the
transition of the task container status from `running` to `terminated`
is not a reliable solution and occasionally the informer misses completion
events.

In order to solve this issue this change would:
* Refactor the informer to reconciler (as analysing pods, deleting jobs,
sending notifications is way too much to do in an informer)
* Make the reporter only care about the current pod state
* Send the completion notification when:
  a) the task pod is in `terminated` state, and
  b) when the related job still exists
Upon sending the notification the reporter deletes the job. The job is
deleted immediately but its pods are deleted in the background.
Therefore the reconciler would keep getting pod events for a while but
as the job no longer exists at that point of time, the reconciler would
not send another completion notification.
* Requeue the notification when notifying the cloud controller fails
without deleting the job so that CC notification is retried during the
next reconciliation loop. Despite being desirable behaviour, this
breaks the existing EATs test as we do not have a proper callback
notification varification facility on the cluster yet and the job does
not get deleted. Therefore the task-reporter EAT is temporarily pended.

[#174598044]

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>