Emit a crash event when app exits with 0
LRPs should never exit, so we now emit crash events regardless of the
exit code. We used the chance to rework our crash event emitting logic
and make it more robust, here are some notable changes:

* Do not rely on CrashLoopBackOff to send crash events. We figured out
that when it crashes a container goes through a terminated state quickly
and then enters a waiting state for a longer period, but while it is
waiting it has a non-nil lastTerminationState, which means that it is
waiwaiting to come up after a crash. We decided to emit crash events in
these two cases (terminated and waiting after a termination) and ignore
everything else.
* Crash counts are now calculated as restart count + 1, because the
first time the app crashes it has 0 restarts, but this is still a crash
* In the crash reporter we ignore all pads that are not eirini apps by
checking the source type annotation (should be "APP"). This is a more
reliable check than parsing the statefulset instance number from the pod
name.
* Introducing the capability to expose an app via a K8S service so that
we could send HTTP requests to it.

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>
Co-authored-by: Georgi Dankov <gddankov@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Mario Nitchev <marionitchev@gmail.com>

[finishes #174817357]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>