Remove unused exposeLRP/unexposeLRP helpers
These are not used anymore since we changed the event reporter EATs.
They might still be very useful to make our tests more reliable in the
future (e.g. we could change the LRP tests to desire an LRP, expose it
and then check it responds).

Deleting them to make the linter happy but let's not forget about them!