Get rid of the explicit multi/single ns support option
Now it is assumed that Eirini supports multiple workloads namespaces,
so that the default single workloads namespace use is a corner case of
the standard multins case. In order to achieve this, Eirini components
are greated with cluster-wide read-only permissions (via ClusterRoles)
and namespaced change permissions (regular Roles); the components
themselves do not perform permissions checks, they assumed that this is
already taken care of.

Although Eirini is not going to be configurable to be scoped down to a
single namespace via configuration ymls, integration tests are still
capable to namespace the binaries (via setting the `WorkloadsNamespace`
property in the binaries' configs) so that tests can be run in
namespace isolation in parallel.

[#174679495]

Co-authored-by: Kieron Browne <kbrowne@vmware.com>