Make crash event reporting robust
Previously the reconciler might run for a crash when container state is
running, and crash is in the lastTerminatedState field. This would not
generate a report as would only do it in the waiting state.

We now send a crash report when we have a lastTerminatedState and the
container is either waiting or running, using an annotation on the pod
to avoid resending events.

[#175611453]

Co-authored-by: Kieron Browne <kbrowne@vmware.com>