Split integration tests script into `run` and `build`
The build action of the script would produce Ginkgo test binaries
locally. Once built, they can be executed with the `build` action. This
would be useful for running the tests on a kind cluster where the Eirini
code is bind mounted as IO in such mounts is very slow.

In the case that no action argument is specified, the script would do
whatever it used to do (i.e. just invoke Ginkgo)

[#176012283]