Split inst idx injector into hook registration and service
With no args passed, the binary both registers the webhook and starts
the webhook service (the previous behaviour), for backwards
compatibility.

New args:
- --register-only means only register the webhook
- --execute-only means only run the webhook service

This way, the webhook registration can be run once with a privileged
service account in a job (requires cluster wide mutating webhook
configuration permissions). And the backing web services can be run in a
scalable deployment with a service account without the cluster
permissions.

[#175878489]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>