Move NATS deployment to skaffold script
NATS needs to be deployed for both helmful and helmless testing. When
included in the skaffold yaml, it can cause flakiness with the start of
the routing components, due to NATS not being ready.

To ensure NATS is fully up before eirini routing components are
deployed, we have moved the helm deployment of NATS out of he skaffold
yaml and into the wrapper script.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>