Log task completion time
This flake used to manifest in a different way
[#175739215](https://www.pivotaltracker.com/story/show/175739215), and
after our attempt at fixing it it showed up again. What we think is
happening is that the reconciler/cluster is under heavy load and the
event that will complete the task is also the one which deletes it
(because the TTL is based on the FinishedAt field of the pod state). To
verify this we've added a log to show these two timestamps.

[#176678637]

Co-authored-by: Mario Nitchev <marionitchev@gmail.com>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>