Address flake where task is deleted before test can see it
On an overloaded machine, the task reconciler might only be notified
when the task TTL has expired. In this case the task will be labelled
and deleted in a single pass, and the test will not see a task labelled
as completed.

To avoid this, bump the TTL to something large for this particular test.

[#176678637]

Co-authored-by: Kieron Browne <kbrowne@vmware.com>