Get rid of "opi" references in codebase
There is one exception: the app/task container name in statefulset/job
is still called opi/opi-task. We depend on that name when getting
metrics, crash events, ect. Also we do not know if other people are not
relying on that name, for example in webhooks.

We will change that name when we do the breaking transition from
statefulsets to deployments

[#176498632]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>