Make `cf4k8s-on-kind.sh -e` build local images
`cf4k8s-on-kind.sh -e` now uses `kbld` and `ytt` to build local images
and render the final Eirini YAML. Images are then loaded in `kind` so
that everything stays local and there is no need to push anything.

This means that from now on, if you want to deploy the local version of
`eirini-release` but still using our released images (this was the
previous behaviour of `cf4k8s-on-kind.sh -e`) you will have to checkout
the released version tag on `eirini`.